% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_quantiles}
\alias{add_quantiles}
\title{Adding numbers and calculate quantiles}
\usage{
add_quantiles(x, y)
}
\arguments{
\item{x}{1st parameter (numeric)}

\item{y}{2nd parameter (numeric)}
}
\description{
Adding numbers and calculate quantiles
}
\examples{
add_quantiles(1, 1)
add_quantiles(10, 1)
}
