% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addsquare}
\alias{addsquare}
\title{Adding numbers and squares}
\usage{
addsquare(x, y)
}
\arguments{
\item{x}{1st parameter (numeric)}

\item{y}{2nd parameter (numeric)}
}
\description{
Adding numbers and squares
}
\examples{
addsquare(1, 1)
addsquare(10, 1)
}
